src/app/modules/core/components/navbar/navbar.component.ts
import { Component, OnDestroy, OnInit } from '@angular/core';
import { Subscription } from 'rxjs';

import { AuthService } from '@/services/auth.service';

@Component({
  selector: 'app-navbar',
  templateUrl: './navbar.component.html',
  styleUrl: './navbar.component.scss'
})
export class NavbarComponent implements OnInit, OnDestroy {
  subscriptions = new Subscription();
  title = 'bookstore';
  isConnected = false;

  constructor(private authService: AuthService) {}

  ngOnInit(): void {
    this.subscriptions.add(
      this.authService.connectedEvent$.subscribe((user) => {
        this.isConnected = !!user;
      })
    );
  }

  ngOnDestroy(): void {
    this.subscriptions.unsubscribe();
  }

  handleDisconnect(): void {
    this.authService.disconnect();
  }
}
